Added a new reporter 'kaizen' similar to 'fail'
Kaizen is the principle of continual and incrimental
imporvement. The purpose of the 'kaizen' reporter is to
allow a team to incrementally improve the lint of a project
by failing the if the number of errors increases.

The kaizen requires for configuration a
bookkeeping file. The first time the lint is run with
this reporter it will fail (based on the same criteria
as the 'fail' reporter) and it will save all current
lint issues to the bookkeeping file.

On all subsquent runnings of the lint the project will
pass so long as the number of error/warnigns/info for a
given file does not increase and any reduction in the number
of errors/warnings/info will be saved to the bookkeeping
file.